lab 4 recipes insert button working and recipes listed in local database

+++ android-project/app/src/main/java/com/tasty/recipesapp/entities/RecipeDao.kt
@@ -1,6 +1,10 @@
 package com.tasty.recipesapp.entities
 
-import androidx.room.*
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.Query
+
 
 @Dao
 interface RecipeDao {
@@ -8,7 +12,7 @@ interface RecipeDao {
     suspend fun insertRecipe(recipe: RecipeEntity)
 
     @Query("SELECT * FROM recipe WHERE internalId = :id")
-    suspend fun getRecipeById(id: Int): RecipeEntity?
+    suspend fun getRecipeById(id: Long): RecipeEntity?
 
     @Query("SELECT * FROM recipe")
     suspend fun getAllRecipes(): List<RecipeEntity>
